browser: reorganize & simplify
