+++ browser/src/components/simple/CustomButton.vue
@@ -1,6 +1,5 @@
 <template>
   <button
-    :id="id"
     type="button"
     class="
     px-4 py-1 min-w-[8rem]
@@ -8,18 +7,9 @@
     bg-disco-cyan
     rounded duration-200
     hover:bg-white hover:outline hover:outline-disco-cyan hover:outline-2 hover:text-disco-cyan"
-    @click="click"
   >
     <slot />
   </button>
 </template>
 
-<script lang="ts">
-export default {
-  name: 'CustomButton',
-  props: {
-    click: { default: () => {}, type: Function },
-    id: { default: 'button', type: String }
-  }
-}
-</script>
+<script lang="ts" setup />
