+++ browser/src/components/training/Finished.vue
