+++ browser/src/components/validation/Validation.vue
@@ -119,7 +119,7 @@ import { IndexedDB } from '@/memory'
 import { toaster } from '@/toast'
 import { WebTabularLoader, WebImageLoader } from '@/data_loader'
 import CustomButton from '@/components/simple/CustomButton.vue'
-import ValidationBar from '@/components/validation/ValidationBar.vue'
+import ValidationBar from '@/components/progress_bars/ValidationBar.vue'
 import Data from '@/components/data/Data.vue'
 import Validator from '@/components/validation/Validator.vue'
 import ButtonCard from '@/components/containers/ButtonCard.vue'
