+++ browser/src/main.ts
@@ -1,10 +1,9 @@
-// import devtools from '@vue/devtools';
 import { createApp } from 'vue'
 import VueApexCharts from 'vue3-apexcharts'
 import Toaster from '@meforma/vue-toaster'
 
 import App from '@/components/App.vue'
-import router from '@/router'
+import { router } from '@/router'
 import { store, key } from '@/store'
 import { createCustomI18n } from './locales/i18n'
 
@@ -17,12 +16,6 @@ tf.ready()
   .then(() => console.log(`Loaded ${tf.getBackend()} backend`))
   .catch(console.error)
 
-/* if (
-  process.env.NODE_ENV === 'development' &&
-  process.env.DEV_TOOLS === 'enabled'
-) {
-  devtools.connect('http://localhost', 8080);
-} */
 // create vue app
 const app = createApp(App)
 const i18n = createCustomI18n()
